Add helper functions for partial pipeline parameters
