+++ java/src/test/kotlin/io/spine/protodata/java/style/JavaCodeStyleFormatterMemoryTest.kt
@@ -35,6 +35,7 @@ import io.spine.protodata.params.WorkingDirectory
 import io.spine.protodata.render.SourceFileSet
 import io.spine.protodata.settings.SettingsDirectory
 import io.spine.protodata.style.indentOptions
+import io.spine.protodata.testing.parametersWithSettingsDir
 import io.spine.protodata.util.Format
 import io.spine.type.toJson
 import java.nio.file.Files.readString
@@ -85,9 +86,7 @@ internal class JavaCodeStyleFormatterMemoryTest {
             this.outputDir = outputDir
             val settingsDir = WorkingDirectory(sandbox).settingsDirectory
             writeSettings(settingsDir)
-            val params = PipelineParameters.newBuilder()
-                .setSettings(settingsDir.path.toDirectory())
-                .buildPartial()
+            val params = parametersWithSettingsDir(settingsDir.path)
             copyResource(fileName, inputDir)
 
             Pipeline(
